change package name from laravel-auth to lockminds-auth to avoid default behavour of removing laravel prefix

+++ src/Traits/CustomClaimsAccessTokenTrait.php
@@ -20,7 +20,7 @@ trait CustomClaimsAccessTokenTrait
             ->where('status', 'valid')
             ->first();
 
-        if($otp->status == "valid"){
+        if(!empty($otp) && $otp->status == "valid"){
             Otp::where("id", $this->getUserIdentifier())->delete();
         }
